+++ Tools.h
@@ -12,13 +12,13 @@
 
 void taskQS(task** tasklist, size_t taille);
 
-uint8_t getCmax(task **tasklist, task *taskToAdd);
+uint16_t getCmax(task **tasklist, task *taskToAdd);
 
-uint8_t getJobRestant(task **tasklist, task *taskToAdd);
+uint16_t getJobRestant(task **tasklist, task *taskToAdd);
 
-uint8_t nbTachePasPlacee(task **tasklist);
+uint16_t nbTachePasPlacee(task **tasklist);
 
-uint8_t getResultCmax(task **taskList);
+uint16_t getResultCmax(task **taskList);
 
 task getTaskH3(task **input);
 
@@ -30,7 +30,7 @@ sol_u* lstTaskToSolu(task** taskList);
 
 void printLstTask(task** taskList);
 
-int8_t addTaskToSolU(sol_u* sol, task* t);
+int16_t addTaskToSolU(sol_u* sol, task* t);
 
 sol_u* allocateNewSolU(void);
 
@@ -38,9 +38,9 @@ sol_u* freeSolU(sol_u* sol);
 
 void* freeAll(sol_u* sol, task** tasks);
 
-int8_t printSolutionU(sol_u* sol);
+int16_t printSolutionU(sol_u* sol);
 
-uint8_t searchArrayForIndex(uint8_t* arr, uint8_t size, uint8_t val);
+uint16_t searchArrayForIndex(uint16_t* arr, uint16_t size, uint16_t val);
 
 int cmp(const void* a, const void* b);
 
@@ -54,4 +54,4 @@ int cmpJob(const void* a, const void* b);
 
 void dumpSolutionUToFile(sol_u* sol);
 
-task* getIthTask(sol_u* sol, uint8_t index);
+task* getIthTask(sol_u* sol, uint16_t index);
